refactor: Model, View, Controller 폴더로 분리하다
