+++ src/Model/BridgeGame.js
@@ -1,5 +1,5 @@
-const BridgeMaker = require('./BridgeMaker');
-const BridgeRanDomNumber = require('./BridgeRandomNumberGenerator');
+const BridgeMaker = require('../BridgeMaker');
+const BridgeRanDomNumber = require('../BridgeRandomNumberGenerator');
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
